requirements.txt
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
peft>=0.7.1
trl>=0.7.10
accelerate>=0.25.0
bitsandbytes>=0.41.3
google-generativeai>=0.3.2
python-dotenv>=1.0.0
gradio>=4.12.0
matplotlib>=3.8.2
pandas>=2.1.4
tqdm>=4.66.1
pyyaml>=6.0.1
sentencepiece>=0.1.99
protobuf>=3.20.0